refactor(open-invoice): encapsulate useSRPanelForOpenInvoiceErrors
